+++ src/utils/operandi_utils/oton/oton_converter.py
@@ -6,7 +6,13 @@ class OTONConverter:
     def __init__(self):
         self.ocrd_validator = OCRDValidator()
 
-    def convert_oton(self, input_path: str, output_path: str, environment: str, with_mets_server: bool):
+    def convert_oton(
+        self,
+        input_path: str,
+        output_path: str,
+        environment: str = "apptainer",
+        with_mets_server: bool = True
+    ):
         list_processor_call_arguments = self.ocrd_validator.validate(input_path)
         nf_file_executable = NextflowFileExecutable()
         nf_file_executable.build_parameters(environment=environment, with_mets_server=with_mets_server)
